backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Async Task Queue
celery==5.3.6
redis==5.0.1

# Audio Processing
demucs==4.0.1
torchaudio>=2.0.0
torch>=2.0.0
torchcodec
ffmpeg-python==0.2.0
soundfile==0.12.1
librosa==0.10.1
numpy>=1.24.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-magic==0.4.27

# Storage
boto3==1.34.34  # Para S3 (opcional)

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.26.0

